[jit] split module interface tests
Summary:
move test_module_interface to its own file, no code logic change

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
